+++ projects/ng-date-picker/src/lib/model/active-date.model.ts
@@ -1,0 +1,4 @@
+export interface ActiveDate {
+  current: Date;
+  previous: Date;
+}