+++ projects/ng-date-picker/src/lib/model/calendar-view-data.ts
@@ -1,7 +1,0 @@
-import { DateRange } from '@angular/material/datepicker';
-
-export class CalendarViewData {
-  startDate!: Date;
-  minDate!: Date;
-  selectedDates!: DateRange<Date>;
-}