+++ projects/ng-date-picker/src/lib/model/date-selection-event.model.ts
@@ -1,5 +1,5 @@
 import { DateRange } from '@angular/material/datepicker';
-import { ISelectDateOption } from './select-date-option';
+import { ISelectDateOption } from './select-date-option.model';
 
 export interface SelectedDateEvent {
   range: DateRange<Date> | null;